Progetto/smart_card_reader/requirements.txt
pyscard
requests 
cryptography
python-dotenv